Paths loaded with OmegaConf from utils
Previously, all paths were given explicitly.

Now, the paths are written to the config.yaml file and config.yaml is included in the .gitignore file.

If you want to look at the writing style of the path, you can look at config.template.yaml.

With the help of methods in utils.py and utils.ipynb files, paths are defined in files.